requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
